`binary.Elf` bugfix, add `binary` package tests
