+++ tests/test_binaries.py
@@ -7,10 +7,36 @@ import os
 import hashlib
 import subprocess
 import time
+from manticore.binary import Elf, CGCElf
 
 #logging.basicConfig(filename = "test.log",
 #                format = "%(asctime)s: %(name)s:%(levelname)s: %(message)s",
 #                level = logging.DEBUG)
+
+
+class TestBinaryPackage(unittest.TestCase):
+    _multiprocess_can_split_ = True
+
+    def test_elf(self):
+        filename = os.path.join(os.path.dirname(__file__), 'binaries', 'basic_linux_amd64')
+        f = Elf(filename)
+        self.assertTrue(
+            [(4194304, 823262, 'r x', 'tests/binaries/basic_linux_amd64', 0, 823262),
+             (7118520, 16112, 'rw ', 'tests/binaries/basic_linux_amd64', 827064, 7320)],
+            list(f.maps())
+        )
+        self.assertTrue([('Running', {'EIP': 4196624})], list(f.threads()))
+        self.assertIsNone(f.getInterpreter())
+
+    def test_decree(self):
+        filename = os.path.join(os.path.dirname(__file__), 'binaries', 'cadet_decree_x86')
+        f = CGCElf(filename)
+        self.assertTrue(
+            [(134512640, 1478, 'r x', 'tests/binaries/cadet_decree_x86', 0, 1478)],
+            list(f.maps())
+        )
+        self.assertTrue([('Running', {'EIP': 134513708})], list(f.threads()))
+
 
 class IntegrationTest(unittest.TestCase):
     _multiprocess_can_split_ = True
