+++ tox.ini
@@ -4,7 +4,7 @@ envlist = py3{6,7}
 [testenv]
 deps = 
 	.[dev]
-commands = nosetests
+commands = nosetests tests
 install_command = pip install --no-binary keystone-engine {opts} {packages}
 
 [testenv:pep8]
@@ -20,4 +20,4 @@ exclude = docs/,examples/,scripts/,tests/
 [flake8]
 ignore = E265,E501,F403,F405,E266,E712,F841,E741,E722,E731
 max-line-length = 160
-exclude = .tox,.*.egg,.git,docs/,examples/,scripts/,tests/
+exclude = .tox,.*.egg,.git,docs/,examples/,scripts/,tests/,parsetab.py,iterpickle.py
